+++ src/main/java/kr/hhplus/be/server/interfaces/orders/api/OrderController.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.orders;
+package kr.hhplus.be.server.interfaces.orders.api;
 
 import org.springframework.web.bind.annotation.*;
 import kr.hhplus.be.server.application.order.OrderFacade;
