+++ src/test/java/kr/hhplus/be/server/support/ControllerTestSupport.java
@@ -9,10 +9,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import kr.hhplus.be.server.application.balance.BalanceFacade;
 import kr.hhplus.be.server.interfaces.balance.BalanceController;
-
+import kr.hhplus.be.server.interfaces.orders.api.OrderController;
 import kr.hhplus.be.server.application.order.OrderFacade;
-import kr.hhplus.be.server.interfaces.orders.OrderController;
-
 import kr.hhplus.be.server.application.product.ProductFacade;
 import kr.hhplus.be.server.interfaces.products.ProductController;
